Fixed Seperated Consolidation-Printing
Consolidated-Branches were marked as using the machine "Consolidated Branch" because the reference to the Recipe was being modified, not a local copy of it. To amend this, I added a property to all recipes, marking if they were consolidated.

This is a temporary fix because one may have multiple recipes being printed.